Changing the input to the create call from Short to Integer

+++ src/main/java/net/floodlightcontroller/core/ImmutablePort.java
@@ -214,7 +214,7 @@ public class ImmutablePort {
              );
     }
 
-    public static ImmutablePort create(String name, Short portNumber) {
+    public static ImmutablePort create(String name, Integer portNumber) {
         return new ImmutablePort(portNumber,
                                          new byte[] { 0, 0, 0, 0, 0, 0 },
                                          name,
